[FEAT] 파일 업로드 구현 (#23)
* chore: gitignore, 파일 업로드 경로 추가

* feat: 파일 업로드/삭제/다운로드 구현

* feat: Channel, ChannelUser 엔티티 추가

* refactor: DTO, Entity 변환을 fromEntity, toEntity로 대체

* refactor: BaseResponseDto로 변경, 복합키 구조 변경

* refactor: 에러 BaseResponseDto 추가

* refactor: RoomUser hasUnread로 변경 및 에러처리 추가

* refactor: File 에러처리 추가

* refactor: 주석 추가, FileEntity -> File로 변경

+++ src/main/java/com/yuguanzhang/lumi/chat/service/ChatServiceImpl.java
@@ -9,9 +9,10 @@ import com.yuguanzhang.lumi.chat.entity.RoomUser;
 import com.yuguanzhang.lumi.chat.repository.ChatRepository;
 import com.yuguanzhang.lumi.chat.repository.RoomRepository;
 import com.yuguanzhang.lumi.chat.repository.RoomUserRepository;
+import com.yuguanzhang.lumi.common.exception.GlobalException;
+import com.yuguanzhang.lumi.common.exception.message.ExceptionMessage;
 import com.yuguanzhang.lumi.user.entity.User;
 import com.yuguanzhang.lumi.user.repository.UserRepository;
-import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -19,7 +20,6 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.NoSuchElementException;
 import java.util.UUID;
 
 @Service
@@ -36,28 +36,24 @@ public class ChatServiceImpl implements ChatService {
     public List<ChatRoomsResponseDto> getChatRooms(UUID userId) {
 
         // 사용자가 속한 방 목록 조회
-        List<RoomUser> roomUsers = roomUserRepository.findByRoomUserId_User_UserId(userId);
+        List<RoomUser> roomUsers = roomUserRepository.findByRoomUserId_UserId(userId);
 
         List<ChatRoomsResponseDto> result = new ArrayList<>();
 
         for (RoomUser ru : roomUsers) {
-            Room room = ru.getRoomUserId().getRoom();
+            Room room = ru.getRoom();
 
             // 채팅방의 상대
             RoomUser senderRoomUser =
-                    roomUserRepository.findByRoomUserId_Room_RoomIdAndRoomUserId_User_UserIdNot(
-                                    room.getRoomId(), userId)
-                            .orElseThrow(() -> new NoSuchElementException("상대방이 존재하지 않습니다."));
+                    roomUserRepository.findByRoomUserId_RoomIdAndRoomUserId_UserIdNot(
+                                              room.getRoomId(), userId)
+                                      .orElseThrow(() -> new GlobalException(
+                                              ExceptionMessage.ROOM_USER_NOT_FOUND));
 
             // 상대방 정보
-            User sender = senderRoomUser.getRoomUserId().getUser();
+            User sender = senderRoomUser.getUser();
 
-            ChatRoomsResponseDto dto =
-                    ChatRoomsResponseDto.builder().roomId(room.getRoomId()).roomName(room.getName())
-                            .opponentId(sender.getUserId()).opponentName(sender.getName())
-                            .lastMessageTime(ru.getLastMessageTime())
-                            .lastMessage(ru.getLastMessageContent())
-                            .unreadMessageCount(ru.getUnreadCount()).build();
+            ChatRoomsResponseDto dto = ChatRoomsResponseDto.fromEntity(room, sender, ru);
 
             result.add(dto);
         }
@@ -71,26 +67,24 @@ public class ChatServiceImpl implements ChatService {
         List<Chat> chatList = chatRepository.findByRoom_RoomId(roomId);
 
         chatList.forEach(chat -> {
-            if (Boolean.FALSE.equals(chat.getIsRead()) && !chat.getUser().getUserId()
-                    .equals(userId)) {
+            if (Boolean.FALSE.equals(chat.getIsRead()) && !chat.getUser()
+                                                               .getUserId()
+                                                               .equals(userId)) {
                 chat.updateIsRead();
             }
         });
 
-        // RoomUser unreadCount 초기화
+        // RoomUser hasUnread 초기화
         RoomUser roomUser =
-                roomUserRepository.findByRoomUserId_Room_RoomIdAndRoomUserId_User_UserId(roomId,
-                                userId)
-                        .orElseThrow(() -> new NoSuchElementException("해당 채팅방에 사용자를 찾을 수 없습니다."));
-        roomUser.resetUnreadCount();
+                roomUserRepository.findByRoomUserId_RoomIdAndRoomUserId_UserId(roomId, userId)
+                                  .orElseThrow(() -> new GlobalException(
+                                          ExceptionMessage.ROOM_NOT_FOUND));
+        roomUser.resetUnread();
         roomUserRepository.save(roomUser);
 
         return chatList.stream()
-                .map(chat -> ChatsResponseDto.builder().roomId(chat.getRoom().getRoomId())
-                        .chatId(chat.getChatId()).messageType(chat.getMessageType())
-                        .senderId(chat.getUser().getUserId()).senderName(chat.getUser().getName())
-                        .message(chat.getContent()).createdAt(chat.getCreatedAt()).build())
-                .toList();
+                       .map(ChatsResponseDto::fromEntity)
+                       .toList();
     }
 
     @Override
@@ -98,46 +92,84 @@ public class ChatServiceImpl implements ChatService {
     public void deleteChat(UUID userId, Long roomId, Long chatId) {
         // 채팅 메세지 조회 (에러처리 필요)
         Chat chat = chatRepository.findByRoom_RoomIdAndChatId(roomId, chatId)
-                .orElseThrow(() -> new EntityNotFoundException("해당 채팅 메세지를 찾을 수 없습니다."));
+                                  .orElseThrow(() -> new GlobalException(
+                                          ExceptionMessage.CHAT_NOT_FOUND));
 
-        // 커스텀 에러 처리 필요
-
-
-        // 삭제했을 때의 마지막 메세지 및 안 읽은 메세지 개수 업데이트 필요
+        if (!chat.getUser()
+                 .getUserId()
+                 .equals(userId)) {
+            throw new GlobalException(ExceptionMessage.UNAUTHORIZED_CHAT_DELETE);
+        }
 
         chatRepository.delete(chat);
+
+        // 삭제했을 때의 마지막 메세지 및 읽음 여부 업데이트 필요
+        Chat lastChat = chatRepository.findTopByRoom_RoomIdOrderByCreatedAtDesc(roomId)
+                                      .orElse(null);
+
+        List<RoomUser> roomUsers = roomUserRepository.findByRoomUserId_RoomId(roomId);
+
+        for (RoomUser ru : roomUsers) {
+            if (lastChat != null) {
+                // 마지막 메세지 갱싱
+                ru.updateLastMessage(lastChat.getContent(), lastChat.getCreatedAt());
+
+                boolean hasUnread = !lastChat.getUser()
+                                             .getUserId()
+                                             .equals(ru.getUser()
+                                                       .getUserId()) && Boolean.FALSE.equals(
+                        lastChat.getIsRead());
+
+
+                if (hasUnread) {
+                    ru.makeUnread();
+                } else {
+                    ru.resetUnread();
+                }
+            } else {
+                // 채팅방에 메세지가 더 이상 없는 경우
+                ru.updateLastMessage(null, null);
+                ru.resetUnread();
+            }
+
+            roomUserRepository.save(ru);
+        }
     }
 
     @Override
     @Transactional
     public RoomUser postChat(ChatRequestDto chatRequestDto, UUID userId, Long roomId) {
         User user = userRepository.findById(userId)
-                .orElseThrow(() -> new IllegalArgumentException("사용자 정보를 찾을 수 없습니다."));
-        Room room = roomRepository.findById(roomId)
-                .orElseThrow(() -> new IllegalArgumentException("채팅방 정보를 찾을 수 없습니다."));
+                                  .orElseThrow(() -> new GlobalException(
+                                          ExceptionMessage.USER_NOT_FOUND));
 
-        Chat chat = Chat.builder().room(room).user(user).content(chatRequestDto.getMessage())
-                .messageType(chatRequestDto.getMessageType()).isRead(false).build();
+        Room room = roomRepository.findById(roomId)
+                                  .orElseThrow(() -> new GlobalException(
+                                          ExceptionMessage.ROOM_NOT_FOUND));
+
+        Chat chat = chatRequestDto.toEntity(room, user);
 
         chatRepository.save(chat);
 
         log.info("chat: {}", chat);
 
-        // 받는 사람 RoomUser의 lastMessage와 안읽은 메세지 수 업데이트
+        // 받는 사람 RoomUser의 lastMessage와 안읽은 메세지, 읽음 여부 업데이트
         RoomUser receiverRoomUser =
-                roomUserRepository.findByRoomUserId_Room_RoomIdAndRoomUserId_User_UserIdNot(roomId,
-                        userId).orElseThrow(() -> new NoSuchElementException("상대방이 존재하지 않습니다."));
+                roomUserRepository.findByRoomUserId_RoomIdAndRoomUserId_UserIdNot(roomId, userId)
+                                  .orElseThrow(() -> new GlobalException(
+                                          ExceptionMessage.ROOM_USER_NOT_FOUND));
+
 
         receiverRoomUser.updateLastMessage(chat.getContent(), chat.getCreatedAt());
-        receiverRoomUser.increaseUnreadCount();
-        roomUserRepository.save(receiverRoomUser);
+        receiverRoomUser.makeUnread();
 
-        // 보낸 사람 RoomUser의 lastMessage 업데이트
+        // 보낸 사람 RoomUser의 lastMessage, 읽음 여부 업데이트
         RoomUser senderRoomUser =
-                roomUserRepository.findByRoomUserId_Room_RoomIdAndRoomUserId_User_UserId(roomId,
-                        userId).orElseThrow();
+                roomUserRepository.findByRoomUserId_RoomIdAndRoomUserId_UserId(roomId, userId)
+                                  .orElseThrow(() -> new GlobalException(
+                                          ExceptionMessage.ROOM_USER_NOT_FOUND));
         senderRoomUser.updateLastMessage(chat.getContent(), chat.getCreatedAt());
-        roomUserRepository.save(senderRoomUser);
+        senderRoomUser.resetUnread();
 
         return receiverRoomUser;
     }
